requirements.txt
google-genai>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

